+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## 0.2.0.0 (unreleased)
 
 * Export version number in Passman.Core.Version
+* Make `passlistPath` non-optional in Config
 
 ## 0.1.0.0